Rearrange view to remove beam status HTML page. Now just show stats, news and graphs
