+++ base/uk.ac.stfc.isis.ibex.ui.beamstatus/plugin.xml
@@ -54,7 +54,7 @@
                id="uk.ac.stfc.isis.ibex.ui.beamstatus.views.BeamStatusView"
                minimized="false"
                moveable="false"
-               ratio="0.7f"
+               ratio="1.0f"
                relationship="right"
                relative="uk.ac.stfc.isis.ibex.ui.perspectives.PerspectiveSwitcher"
                showTitle="false"
@@ -66,7 +66,7 @@
                id="uk.ac.stfc.isis.ibex.ui.beamstatus.views.BeamStatusHourlyGraphView"
                minimized="false"
                moveable="false"
-               ratio="0.5f"
+               ratio="0.6f"
                relationship="right"
                relative="uk.ac.stfc.isis.ibex.ui.beamstatus.views.BeamStatusView"
                showTitle="false"
